#50 - Implement route resolvers
- a route resolver follow a similar mechanism to response transports. A resolver
receives the router object which gives it access to the response and can resolve
the request, if the resolver returns not FALSE the resolver queue will be halted
and the resolved route returned by the router.

- The resolver also implements reverse routing through the generate method. Here
the resolver will receive the path and the query to generate a route for. if
the resolver returns not FALSE the resolver queue will be halted and the generated
route returned by the router.