#146 - Implement snippet helper
Helper defines two methods a define() method to create a snippet and an
expand() method to render a snippet.

A snipppet is expanded using heredoc syntax. Heredoc text behaves just
like a double-quoted string, without the double quotes. This means that
quotes in a heredoc do not need to be escaped.