#167 -  Refactor router
- Replace the resolver queue with a two-step resolution mechanism. The
framework will resolve the component from the route, and the component
will resolve the rest of the route.

- Router exposes 4 key methods: compile, resolve, generate and qualify.
The compile method is responsible to generate a route object from the
provided parameters, the route object is used to resolve or generate a
route and in a final step the qualify method will transform the route to
a url string.

- Move the site, page and redirect resolving into dispatcher behaviors,
and decouple resolvers from the request of the router.