#167 -  Improve routing API
- Add getResolver() method to ComPagesDispatcherRouterRoute to allow the
route to return a resolver identifier based on the information it contains.

- Allow to have routes without a host information, they will be processed
by the default ComPagesDispatcherResolver. This would be the default for
simple components that only need a single resolver.

- Rename compile() method to getRoute(). Do not let the abstract router
call this method during resolve or generate, make this the responsability
of the implementation.

- Pass in the request query as routing parameters when resolving the page
route, required to be able to handle pagination.