#188 - Refactor validatable behavior
- Move from controller to dispatcher to make sure validation can happen
as early as possible

- Add request sanitization support through santizeRequest() method. Any
data that is not in the whitelist or the schema is removed. By default
the whitelist exists of '_method', '_action', '_format'

- Handle schema constraints for PUT and POST request. A PUT request always
requires a complete schema representation to be present in the request. A
POST request check if the data attribute is required, if the constraint
are not met a 400 BAD REQUEST will be returned.