#176 - Implement filterable behavior
This behavior adsd a 'filter' state in accoordance with the JSONAOPI 1.0
specification.

The key should be the attribute to filter on, while the value reflect
what to filter for.

Multiple comma seperated values result in an implict OR while multiple
filters result in an implicit AND.

- OR: ?filter[foo]=bar,cube
- AND: ?filter[foo]=bar&filter[name]=joe

For queries that need comparisons other than simple equals, operators
are supported for inequality, greater-than, greater-than-or-equal,
less-than, and less-than-or-equal-to.

In order, the operators are: eg, neq, gt, gte, lt, and lte. Operators must
be followed by a colon. For example  ?filter[size]=gt:10