#214 - Implement http cache proxy as anonymous function
The proxy offers Cache Validation using ETag and Last-Mofified and Cache
Expiration using the `max_age`function parameter. Default is FALSE.

If the resource was served from cache the proxy will set the Age header
with the calculated the response was generated or validated by the origin
server. After succesfull validation the proxy will return a 304 Not Modified
together with the current date in a Date header to allow clients to
freshen their own stored response.