#212 - Add failed revalidation support
If revalidation of the cache fail due to a http error and the debug is
not enabled or the cache must not be revalidated serve a stale response
from the cache.

The implementation will add a Warning: 111 Revalidation Failed header in
accoordance with https://tools.ietf.org/html/rfc7234#section-5.5.2 and
it will set the Cache-Status header to STALE to indicate the response is
stale and was served from cache after revalidation failed.

If debug is enabled a correspponding http exception will be thrown and
a stale response will not be returned.