#166 - Add casting methods
Add toJosn(), toXml() and toHtml() method to the data object. This allows
to cast a data object to each format.

The toXml() and toHtml() methods return a DomDocumentXml or DomDocumentHtml
object that gives easy access to xpath expressions through the query() and
evaluate() methods. It can also be casted to a string.