#910 - Improve jsonapi links routing
Add format specific links if the page is not 'json' to allow discovering and crawling content
through the jsonapi. The jsonapi entity links are generated using the collection primary keys.

A collection can define an additional route property, which allows to set the route to be
used to route the collection entities. If the collection route is false, no entity links
will be included in the jsonapi